+++ packages/pinia-orm/build.config.ts
@@ -6,11 +6,7 @@ export default defineBuildConfig({
     // bundling
     'src/index',
     'src/decorators',
-    // bundleless, or just copy assets
-    // { input: 'src/model/decorators/attributes/types/', outDir: 'dist/decorators' },
-    // { input: 'src/model/decorators/attributes/relations/', outDir: 'dist/decorators' },
-    // { input: 'src/model/decorators/Cast', outDir: 'dist/decorators', builder: 'mkdist', name: 'Cast' },
-    // { input: 'src/model/decorators/Mutate.ts', outDir: 'dist/decorators' },
+    'src/helpers',
   ],
   declaration: true,
   clean: true,
