+++ packages/pinia-orm/src/helpers.ts
@@ -1,0 +1,1 @@
+export * from './composables/mapRepos'
