+++ packages/pinia-orm/src/index.ts
@@ -6,23 +6,6 @@ export * from './store/Store'
 export * from './database/Database'
 export * from './schema/Schema'
 export * from './model/Model'
-export * from './model/decorators/attributes/types/Attr'
-export * from './model/decorators/attributes/types/Str'
-export * from './model/decorators/attributes/types/Num'
-export * from './model/decorators/attributes/types/Bool'
-export * from './model/decorators/attributes/types/Uid'
-export * from './model/decorators/attributes/relations/HasOne'
-export * from './model/decorators/attributes/relations/BelongsTo'
-export * from './model/decorators/attributes/relations/BelongsToMany'
-export * from './model/decorators/attributes/relations/HasMany'
-export * from './model/decorators/attributes/relations/HasManyBy'
-export * from './model/decorators/attributes/relations/MorphOne'
-export * from './model/decorators/attributes/relations/MorphTo'
-export * from './model/decorators/attributes/relations/MorphMany'
-export * from './model/decorators/Contracts'
-export * from './model/decorators/Cast'
-export * from './model/decorators/Mutate'
-export * from './model/decorators/NonEnumerable'
 export * from './model/attributes/Attribute'
 export * from './model/casts/CastAttribute'
 export * from './model/attributes/types/Type'
@@ -30,5 +13,4 @@ export * from './repository/Repository'
 export * from './interpreter/Interpreter'
 export * from './query/Query'
 export * from './query/Options'
-export * from './helpers/Helpers'
 export * from './plugin/Plugin'
