+++ packages/pinia-orm/tests/unit/model/Model_Attrs_UID.spec.ts
@@ -1,6 +1,7 @@
 import { describe, expect, it } from 'vitest'
 
-import { Model, Uid } from '../../../src'
+import { Model} from '../../../src'
+import { Uid } from '../../../src/decorators'
 import { mockUid } from '../../helpers'
 
 describe('unit/model/Model_Attrs_UID', () => {
