+++ packages/pinia-orm/tests/unit/model/Model_Casts_Array.spec.ts
@@ -1,6 +1,7 @@
 import { beforeEach, describe, expect, it } from 'vitest'
 
-import { Attr, Cast, Model, useRepo } from '../../../src'
+import { Model, useRepo } from '../../../src'
+import { Attr, Cast } from '../../../src/decorators'
 import { assertState } from '../../helpers'
 
 describe('unit/model/Model_Casts_Array', () => {
