+++ packages/pinia-orm/tests/unit/model/Model_Keys.spec.ts
@@ -1,6 +1,7 @@
 import { beforeEach, describe, expect, it } from 'vitest'
 
-import { Attr, Model } from '../../../src'
+import { Model } from '../../../src'
+import { Attr } from '../../../src/decorators'
 
 describe('unit/model/Model_Keys', () => {
   beforeEach(() => {
