+++ packages/pinia-orm/tests/unit/model/Model_STI.spec.ts
@@ -1,6 +1,7 @@
 import { beforeEach, describe, expect, it } from 'vitest'
 
-import { Attr, Model, useRepo } from '../../../src'
+import { Model, useRepo } from '../../../src'
+import { Attr } from '../../../src/decorators'
 import { assertState } from '../../helpers'
 
 describe('unit/model/Model_STI', () => {
